Added description for all methods & modules
