+++ README.md
@@ -1,1 +1,10 @@
-Distance Calculator Web Application
+## Distance Calculator Web Application
+
+## Application Contains Backend REST API for Distance Calculator 
+
+## API endpoint
+To get the data(qurey history) from DB use /history
+example: http://localhost:8080/history
+
+To post new query into the DB /history/new
+example: http://localhost:8080/history/new
