+++ geoCode.js
@@ -1,6 +1,6 @@
 const axios = require('axios');
 
-
+//Method for retrieving GeoCodes using Nomiinatim API
 const getGeocode = async (address) => {
     try {
         const response = await axios.get('https://nominatim.openstreetmap.org/search', {
@@ -18,7 +18,7 @@ const getGeocode = async (address) => {
         const { lat, lon } = response.data[0];
         return { latitude: parseFloat(lat), longitude: parseFloat(lon) };
     } catch (error) {
-        throw new Error(`Geocoding failed for "${address}": ${error.message}`);
+        throw new Error(error.message);
     }
 };
